Rename class files
As discussed with moorscode.

Adjusts the file names to be in line with the rules as laid out in the `Yoast.Files.FileName` sniff.
Ref: Yoast/yoastcs 48

* Remove `class-` prefixes from the plugin file names.
* Make sure that the file names accurately reflect the class names (excluding the prefixes).

Where necessary, adjusted any `include`/`require`s to point to the new file names.
